requirements.txt
pynput
sounddevice
soundfile
openai
Pillow
pyperclip
pyautogui
pystray
numpy
